add bot admin validation

+++ handlers/delete.py
@@ -4,12 +4,13 @@ from utils.chatMember import is_admin
 
 
 async def delete_message(message: types.Message):
+    if not await is_admin(message.chat, await message.bot.get_me()):
+        return
+
     if not await is_admin(message.chat, message.from_user):
-        await message.reply("You are not an admin of this group.")
         return
 
     if not message.reply_to_message:
-        await message.reply("Please reply to the message you want to delete.")
         return
 
     await message.delete()
